Revert "Remove deprecated ls and cat commands (#142)"
This reverts commit d3a805c09f55d25a1cc64062b82b5bda2213677d.
